pyproject.toml
[project]
name = "site"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flet[all]>=0.28.3",
    "flet-cli>=0.28.3",
    "python-dotenv>=1.1.0",
]


[tool.flet.app]
path = "src"
